refactor(front): Extract conversation page components for SPA migration
